requirements.txt
google-genai>=1.0.0
fastapi
chromadb
uvicorn
watchdog
streamlit
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
pywebview
plyer
cryptography
